+++ lib/aixm/component/helipad.rb
@@ -12,8 +12,7 @@ module AIXM
     #     xy = AIXM.xy
     #   )
     #   helipad.z = AIXM.z or nil
-    #   helipad.length = AIXM.d or nil   # must use same unit as width
-    #   helipad.width = AIXM.d or nil    # must use same unit as length
+    #   helipad.dimensions = AIXM.r or nil
     #   helipad.surface = AIXM.surface
     #   helipad.marking = String or nil
     #   helipad.fato = AIXM.fato or nil
@@ -77,11 +76,8 @@ module AIXM
       # @return [AIXM::Z, nil] elevation in +:qnh+
       attr_reader :z
 
-      # @return [AIXM::D, nil] length
-      attr_reader :length
-
-      # @return [AIXM::D, nil] width
-      attr_reader :width
+      # @return [AIXM::R, nil] dimensions
+      attr_reader :dimensions
 
       # @return [String, nil] markings
       attr_reader :marking
@@ -120,20 +116,9 @@ module AIXM
         @z = value
       end
 
-      def length=(value)
-        @length = if value
-          fail(ArgumentError, "invalid length") unless value.is_a?(AIXM::D) && value.dist > 0
-          fail(ArgumentError, "invalid length unit") if width && width.unit != value.unit
-          @length = value
-        end
-      end
-
-      def width=(value)
-        @width = if value
-          fail(ArgumentError, "invalid width") unless value.is_a?(AIXM::D)  && value.dist > 0
-          fail(ArgumentError, "invalid width unit") if length && length.unit != value.unit
-          @width = value
-        end
+      def dimensions=(value)
+        fail(ArgumentError, "invalid dimensions") unless value.nil? || value.is_a?(AIXM::R)
+        @dimensions = value
       end
 
       def marking=(value)
@@ -175,10 +160,11 @@ module AIXM
             tla.valElev(z.alt)
             tla.uomDistVer(z.unit.upcase.to_s)
           end
-          tla.valLen(length.dist.trim) if length
-          tla.valWid(width.dist.trim) if width
-          tla.uomDim(length.unit.to_s.upcase) if length
-          tla.uomDim(width.unit.to_s.upcase) if width && !length
+          if dimensions
+            tla.valLen(dimensions.length.to_m.dim.trim)
+            tla.valWid(dimensions.width.to_m.dim.trim)
+            tla.uomDim('M')
+          end
           unless  (xml = surface.to_xml).empty?
             tla << xml.indent(2)
           end