+++ spec/lib/aixm/component/runway_spec.rb
@@ -32,23 +32,13 @@ describe AIXM::Component::Runway do
     end
   end
 
-  describe :length= do
+  describe :dimensions= do
     it "fails on invalid values" do
-      _([:foobar, 0, 1, AIXM.d(0, :m)]).wont_be_written_to subject, :length
+      _([:foobar, 0, 1, AIXM.d(0, :m)]).wont_be_written_to subject, :dimensions
     end
 
     it "accepts nil value" do
-      _([nil]).must_be_written_to subject, :length
-    end
-  end
-
-  describe :width= do
-    it "fails on invalid values" do
-      _([:foobar, 0, 1, AIXM.d(0, :m)]).wont_be_written_to subject, :width
-    end
-
-    it "accepts nil value" do
-      _([nil]).must_be_written_to subject, :width
+      _([nil]).must_be_written_to subject, :dimensions
     end
   end
 
@@ -218,7 +208,7 @@ describe AIXM::Component::Runway do
 
     it "builds correct minimal OFMX" do
       AIXM.ofmx!
-      %i(length width status remarks).each { subject.send(:"#{_1}=", nil) }
+      %i(dimensions status remarks).each { subject.send(:"#{_1}=", nil) }
       %i(composition preparation condition pcn siwl_weight siwl_tire_pressure auw_weight remarks).each { subject.surface.send(:"#{_1}=", nil) }
       %i(forth back).each { subject.send(_1).instance_eval { @lightings.clear } }
       _(subject.to_xml).must_equal <<~END